# api/http/handler.go
package http

import (
	"encoding/json"
	"net/http"

	"github.com/vinod901/opendq-go/internal/policy"
	"github.com/vinod901/opendq-go/internal/tenant"
	"github.com/vinod901/opendq-go/internal/workflow"
)

// Handler holds HTTP handlers
type Handler struct {
	tenantManager   *tenant.Manager
	policyManager   *policy.Manager
	workflowEngine  *workflow.Engine
}

// NewHandler creates a new HTTP handler
func NewHandler(
	tenantManager *tenant.Manager,
	policyManager *policy.Manager,
	workflowEngine *workflow.Engine,
) *Handler {
	return &Handler{
		tenantManager:  tenantManager,
		policyManager:  policyManager,
		workflowEngine: workflowEngine,
	}
}

// RegisterRoutes registers all HTTP routes
func (h *Handler) RegisterRoutes(mux *http.ServeMux) {
	// Health check
	mux.HandleFunc("/health", h.HealthCheck)

	// API documentation
	mux.HandleFunc("/api/docs", h.serveOpenAPISpec)
	mux.HandleFunc("/api/docs/", h.serveSwaggerUI)

	// Tenant routes
	mux.HandleFunc("/api/v1/tenants", h.handleTenants)
	mux.HandleFunc("/api/v1/tenants/", h.handleTenant)

	// Policy routes
	mux.HandleFunc("/api/v1/policies", h.handlePolicies)
	mux.HandleFunc("/api/v1/policies/", h.handlePolicy)

	// Workflow routes
	mux.HandleFunc("/api/v1/workflows", h.handleWorkflows)
	mux.HandleFunc("/api/v1/workflows/", h.handleWorkflow)

	// Lineage routes
	mux.HandleFunc("/api/v1/lineage", h.handleLineage)
}

// serveOpenAPISpec serves the OpenAPI specification
func (h *Handler) serveOpenAPISpec(w http.ResponseWriter, r *http.Request) {
	http.ServeFile(w, r, "api/openapi.yaml")
}

// serveSwaggerUI serves a simple Swagger UI redirect
func (h *Handler) serveSwaggerUI(w http.ResponseWriter, r *http.Request) {
	// Redirect to Swagger UI hosted version with our spec
	swaggerURL := "https://petstore.swagger.io/?url=http://localhost:8080/api/docs"
	w.Header().Set("Content-Type", "text/html")
	w.Write([]byte(`<!DOCTYPE html>
<html>
<head>
    <title>OpenDQ API Documentation</title>
    <meta http-equiv="refresh" content="0; url='` + swaggerURL + `'" />
</head>
<body>
    <p>Redirecting to <a href="` + swaggerURL + `">Swagger UI</a>...</p>
    <p>Alternatively, you can:</p>
    <ul>
        <li>View the <a href="/api/docs">OpenAPI Spec (YAML)</a></li>
        <li>Use <a href="https://editor.swagger.io/?url=http://localhost:8080/api/docs">Swagger Editor</a></li>
    </ul>
</body>
</html>`))
}

// HealthCheck handles health check requests
func (h *Handler) HealthCheck(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"status": "healthy",
	})
}

// Tenant handlers

func (h *Handler) handleTenants(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		h.listTenants(w, r)
	case http.MethodPost:
		h.createTenant(w, r)
	default:
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
	}
}

func (h *Handler) handleTenant(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		h.getTenant(w, r)
	case http.MethodPut:
		h.updateTenant(w, r)
	case http.MethodDelete:
		h.deleteTenant(w, r)
	default:
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
	}
}

func (h *Handler) listTenants(w http.ResponseWriter, r *http.Request) {
	tenants, err := h.tenantManager.ListTenants(r.Context())
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(tenants)
}

func (h *Handler) createTenant(w http.ResponseWriter, r *http.Request) {
	var req struct {
		Name     string                 `json:"name"`
		Slug     string                 `json:"slug"`
		Metadata map[string]interface{} `json:"metadata"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	tenant, err := h.tenantManager.CreateTenant(r.Context(), req.Name, req.Slug, req.Metadata)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(tenant)
}

func (h *Handler) getTenant(w http.ResponseWriter, r *http.Request) {
	// Extract tenant ID from path
	// Implementation depends on routing library
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *Handler) updateTenant(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *Handler) deleteTenant(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

// Policy handlers

func (h *Handler) handlePolicies(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		h.listPolicies(w, r)
	case http.MethodPost:
		h.createPolicy(w, r)
	default:
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
	}
}

func (h *Handler) handlePolicy(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		h.getPolicy(w, r)
	case http.MethodPut:
		h.updatePolicy(w, r)
	case http.MethodDelete:
		h.deletePolicy(w, r)
	default:
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
	}
}

func (h *Handler) listPolicies(w http.ResponseWriter, r *http.Request) {
	tenantID := r.URL.Query().Get("tenant_id")
	policies, err := h.policyManager.ListPolicies(r.Context(), tenantID)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(policies)
}

func (h *Handler) createPolicy(w http.ResponseWriter, r *http.Request) {
	var pol policy.Policy
	if err := json.NewDecoder(r.Body).Decode(&pol); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	if err := h.policyManager.CreatePolicy(r.Context(), &pol); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(pol)
}

func (h *Handler) getPolicy(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *Handler) updatePolicy(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *Handler) deletePolicy(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

// Workflow handlers

func (h *Handler) handleWorkflows(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		h.listWorkflows(w, r)
	case http.MethodPost:
		h.createWorkflow(w, r)
	default:
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
	}
}

func (h *Handler) handleWorkflow(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		h.getWorkflow(w, r)
	case http.MethodPost:
		h.transitionWorkflow(w, r)
	default:
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
	}
}

func (h *Handler) listWorkflows(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *Handler) createWorkflow(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *Handler) getWorkflow(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *Handler) transitionWorkflow(w http.ResponseWriter, r *http.Request) {
	var req struct {
		WorkflowName string `json:"workflow_name"`
		Event        string `json:"event"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	if err := h.workflowEngine.Transition(r.Context(), req.WorkflowName, req.Event); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	state, _ := h.workflowEngine.GetCurrentState(req.WorkflowName)

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"status":        "success",
		"current_state": state,
	})
}

// Lineage handlers

func (h *Handler) handleLineage(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		h.getLineage(w, r)
	case http.MethodPost:
		h.createLineageEvent(w, r)
	default:
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
	}
}

func (h *Handler) getLineage(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *Handler) createLineageEvent(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}