# api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenDQ API
  description: |
    OpenDQ is an enterprise-grade data quality platform. This API provides 
    endpoints for managing datasources, data quality checks, schedules, 
    alerts, and views.
    
    ## Authentication
    All API endpoints (except /health) require authentication via Bearer token.
    
    ## Multi-Tenancy
    For multi-tenant deployments, include the `X-Tenant` header with your tenant slug.
  version: 1.0.0
  contact:
    name: OpenDQ Team
    url: https://github.com/vinod901/opendq-go
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
paths:
...
  /tenants:
    get:
      tags: [Tenants]
      summary: List tenants
      operationId: listTenants
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tenant'
    post:
      tags: [Tenants]
      summary: Create tenant
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        metadata:
          type: object
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTenantRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        metadata:
          type: object
...
security:
  - BearerAuth: []